Make SubActor trait and behavior separate from ResourceCrate
