+++ OpenRA.Mods.OpenE2140/Traits/Resources/ResourceCrate.cs
@@ -12,146 +12,31 @@
 #endregion
 
 using JetBrains.Annotations;
-using OpenRA.Mods.Common;
-using OpenRA.Mods.Common.Traits;
+using OpenRA.Mods.OpenE2140.Traits.SubActors;
 using OpenRA.Traits;
 
 namespace OpenRA.Mods.OpenE2140.Traits.Resources;
 
-public interface ISubActorParent
-{
-	WVec SubActorOffset { get; }
-}
-
 [UsedImplicitly]
 [Desc("This actor is a resource crate.")]
-public class ResourceCrateInfo : TraitInfo, IFacingInfo
+public class ResourceCrateInfo : TraitInfo, Requires<SubActorInfo>
 {
-	public WAngle GetInitialFacing() => WAngle.Zero;
-
 	public override object Create(ActorInitializer init)
 	{
-		return new ResourceCrate(init);
+		return new ResourceCrate(init.Self);
 	}
 }
 
-public class ResourceCrate : IFacing, IOccupySpace, ITick
+public class ResourceCrate
 {
-	private Actor? parentActor;
-	private ISubActorParent? subActorParent;
-	private IFacing? parentFacing;
-	private WPos centerPosition;
-	private WRot orientation;
-	private CPos location;
-
 	public readonly Actor Actor;
+	public readonly SubActor SubActor;
+
 	public int Resources;
 
-	public Actor? ParentActor
+	public ResourceCrate(Actor self)
 	{
-		get => this.parentActor;
-		set
-		{
-			if (value == null)
-			{
-				this.Actor.World.ActorMap.AddInfluence(this.Actor, this);
-			}
-			else
-			{
-				this.centerPosition = value.CenterPosition;
-				this.Actor.World.ActorMap.RemoveInfluence(this.Actor, this);
-			}
-
-			this.parentActor = value;
-			this.subActorParent = value?.TraitOrDefault<ISubActorParent>();
-
-			this.parentFacing = value?.TraitOrDefault<IFacing>();
-		}
+		this.Actor = self;
+		this.SubActor = self.Trait<SubActor>();
 	}
-
-	public ResourceCrate(ActorInitializer init)
-	{
-		this.Actor = init.Self;
-
-		var locationInit = init.GetOrDefault<LocationInit>();
-		if (locationInit != null)
-		{
-			this.location = locationInit.Value;
-			this.centerPosition = init.World.Map.CenterOfCell(this.location);
-		}
-
-		var parentInit = init.GetOrDefault<ParentActorInit>();
-		if (parentInit != null)
-		{
-			init.World.AddFrameEndTask(w =>
-			{
-				var parentActor = parentInit.Value.Actor(w).Value;
-				if (this.Actor.IsInWorld && parentActor.IsInWorld)
-					this.ParentActor = parentActor;
-			});
-		}
-
-		this.Facing = init.GetValue<FacingInit, WAngle>(WAngle.Zero);
-	}
-
-	void ITick.Tick(Actor self)
-	{
-		if (this.parentActor == null)
-		{
-			return;
-		}
-
-		this.orientation = this.parentActor.Orientation;
-		this.centerPosition = this.parentActor.CenterPosition;
-		this.location = this.parentActor.Location;
-	}
-
-	#region IFacing
-
-	[Sync]
-	public WAngle Facing
-	{
-		get => this.parentActor?.Orientation.Yaw ?? this.orientation.Yaw;
-		set
-		{
-			if (this.parentActor == null || this.parentFacing == null)
-			{
-				this.orientation = this.orientation.WithYaw(value);
-			}
-			else
-			{
-				this.parentFacing.Facing = value;
-				this.orientation = this.parentFacing.Orientation;
-			}
-		}
-	}
-
-	public WRot Orientation => this.parentActor?.Orientation ?? this.orientation;
-
-	public WAngle TurnSpeed => WAngle.Zero;
-
-	#endregion
-
-	#region IOccupySpace
-
-	[Sync]
-	public WPos CenterPosition
-	{
-		get
-		{
-			if (this.parentActor == null || this.subActorParent == null)
-				return this.centerPosition;
-
-			return this.parentActor.CenterPosition + this.subActorParent.SubActorOffset;
-		}
-	}
-
-	public CPos TopLeft => this.parentActor?.OccupiesSpace?.TopLeft ?? this.Actor.World.Map.CellContaining(this.centerPosition);
-
-	public (CPos, SubCell)[] OccupiedCells()
-	{
-		return this.parentActor?.OccupiesSpace?.OccupiedCells() ?? new[] { (this.location, SubCell.FullCell) };
-	}
-
-	#endregion
 }
