+++ mods/e2140/content/shared/misc/crate/rules.yaml
@@ -11,4 +11,5 @@ crate:
 	Interactable:
 	Targetable:
 		TargetTypes: Ground
+	SubActor:
 	WithFacingSpriteBody:
